requirements.txt
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2021.5.30
chardet==4.0.0
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0a1
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.17.1
pycryptodome==3.10.1
pyrsistent==0.17.3
python-dotenv==0.17.1
requests==2.25.1
rlp==2.0.1
six==1.16.0
toolz==0.11.1
urllib3==1.26.5
varint==1.0.2
web3==5.19.0
websockets==8.1
appdirs==1.4.4
asn1crypto==1.4.0
astroid==2.5.6
base58==2.1.0
bech32==1.2.0
bip32==0.1
black==21.5b2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
coincurve==13.0.0
dacite==1.6.0
ecdsa==0.17.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mnemonic==0.19
mypy-extensions==0.4.3
pathspec==0.8.1
pyband==0.1.0
pycparser==2.20
pylint==2.8.3
regex==2021.4.4
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.5
wrapt==1.12.1